Implement optional persistent registry lookup cache and better 429 handling
This adds a `--cache` flag to builds which will read from a cache file and write back out to it a set of images that are ~safe to cache (and because it's just a file, we can clear cache entries trivially by deleting the relevant entries from the file).

This also adds better 429 behavior that's twofold:

1. cap our maximum registry requests at ~500/min with an initial burst of 100 unless we hit a 429 (which should, in theory, keep us closer to staying under the Hub abuse rate limits)

2. actually retry requests when they give us a 429 response (capping both new requests and retries at the same ~500/min which is ~8/sec)

As a final bonus, I added/generated an appropriate cache file for our local test suite which brings the total number of actual Hub/registry requests our test suite makes down to one single image lookup.