Add a 3 hour timeout to individual builds
This was part of the old system, but we hadn't yet had the problems that require it resurface yet.  The limit here is kind of arbitrary, but the specific value isn't as important as having one -- we had a build hang for 6d (due to some tool that doesn't have a timeout of its own but arguably should), and this prevents that from stalling our whole process.