Fix the cache broken by the ociref changes
One part of the code was storing keys in our `sync.Map` as a `string` and the other as a `Reference` so they never matched. 😭

If `sync.Map` had been written in the brave new world of Go generics, we'd have a compilation error to help us catch this. 🤦

We *could* get clever with something like a `--network=none` container that forces `builds` to use the cache, but it's a bit complex to set up correctly (so instead I think we should wait for the eventual `sync/v2` upstream Go package or fork/write/switch to a separate "generic" / type-safe `sync.Map` package in the future).