Add Tianon's cursed "infosiftr/moby" library file for more useful edge cases
This includes the edge case of arch-particular but not arch-specific tags (similar to busybox, but with the primary difference being that they *share* a sourceId instead of having disparate sourceIds).

In this, I have hand-modified the `all-tags.json` file to what it *should* be so that the tests will fail until it is fixed appropriately (but all other generated files are as-generated).