Switch from the `subset` branch back to `main`
When I created these branches ~a year ago, I had lofty goals that we'd be able to turn on "everything" quickly enough that the `subset` branch would be short-lived and that `main` could stay faithfully tracking the "everything is included" world, but the reality is much messier.

In an attempt to be extremely pragmatic about the reality, my intention is to delete the existing (now extremely outdated) `main` branch and simply rename `subset` to `main`.  It will then organically grow to include "everything" as we've been doing (and eventually we'll even ditch `subset.txt` 🤞).