Update Jenkinsfile.* to use shallow clones of meta
We don't need a full (unshallow) clone for any of these -- the jobs don't look at the commit history and we can make new commits without having history locally.

This will help on architectures that have been struggling to clone the repository since the move (because a "full" clone is now bigger than it was when we started all this), but is also generally better all around.