Adjust deploy input normalization produce valid inputs
The new `deploy --dry-run` feature relies on the assumption that JSON-ified normalized input objects are valid input back, which wasn't an assumption we were validating.

This adjusts `TestNormalizeInput` to take the output of normalization and pass it back through for an added "roundtrip" test.  In doing so, I found a single edge case where our normalized objects were *not* valid/correct inputs (tag->tag copy of a manifest) and made a minor adjustment to the normalization to close that hole.

I also added a new test case for that edge case, and made sure *it* round-trips correctly too.