Fix `deploy --dry-run` deadlock
Once we got over 1000 items in our deploy list, we deadlocked because nothing consumes the channel until we're done with the "queue everything up" loop, so we need to start consuming that list right away instead.  As long as it's part of our waitgroup, this should be fine and work correctly.