Remove `SOURCE_DATE_EPOCH` from all our builds
This was a misguided attempt at some initial aspects of reproducibility.

The actual goal is of dubious value compared to the amount of effort required, but the half-measure is outright incorrect/useless, and a consistent and repeated source of user confusion.

This dramatically simplifies our build commands because all the annotations get set to the same value, so we can drop half the `docker buildx build` arguments.